Move driver state to a new package
