+++ src/py/flwr/server/app.py
@@ -29,9 +29,8 @@ from flwr.common.logger import log
 from flwr.proto.driver_pb2_grpc import add_DriverServicer_to_server
 from flwr.proto.transport_pb2_grpc import add_FlowerServiceServicer_to_server
 from flwr.server.client_manager import ClientManager, SimpleClientManager
-from flwr.server.driver.driver_client_manager import DriverClientManager
 from flwr.server.driver.driver_servicer import DriverServicer
-from flwr.server.driver.state import DriverState
+from flwr.server.grpc_server.driver_client_manager import DriverClientManager
 from flwr.server.grpc_server.flower_service_servicer import FlowerServiceServicer
 from flwr.server.grpc_server.grpc_server import (
     generic_create_grpc_server,
@@ -39,6 +38,7 @@ from flwr.server.grpc_server.grpc_server import (
 )
 from flwr.server.history import History
 from flwr.server.server import Server
+from flwr.server.state.state import DriverState
 from flwr.server.strategy import FedAvg, Strategy
 
 DEFAULT_SERVER_ADDRESS = "[::]:8080"
